Fix issue where tags disappear on second edit
The problem was that "this.tags" is how the image/journal
detail component was communicating its selected tags
back up to the parent component. But that didn't get
populated until the first call to changeTags(). Which
wasn't a problem when we were only adding, but once we
started to edit, you could get a case where the default
tags were present but changeTags() never got called.